1st commit-incomplete program
Need to add some formatting details and comments, When completed colors will display in the correct color, the player will be able to enter a command to list the valid colors and a command to list colors they have already guessed. Considering adding a timer and eventually the game will display in a formatted window.